requirements.txt
Django==4.0
djangorestframework==3.13.0
psycopg2-binary==2.9.2

flake8==4.0.1
isort==5.10.1
